Refactored copySnippet function to be passed in by corresponding cluster page
