+++ code/workspaces/web-app/src/containers/clusters/ClustersContainer.js
@@ -3,15 +3,16 @@ import { useCurrentProjectKey } from '../../hooks/currentProjectHooks';
 import { useCurrentUserPermissions } from '../../hooks/authHooks';
 import ProjectClustersContainer from './ProjectClustersContainer';
 
-const ClustersContainer = ({ clusterType }) => {
+const ClustersContainer = ({ clusterType, copySnippet = undefined }) => {
   const { value: projectKey } = useCurrentProjectKey();
   const { value: userPermissions } = useCurrentUserPermissions();
   const modifyData = true;
-  return ProjectClustersContainer({ clusterType, projectKey, userPermissions, modifyData });
+  return ProjectClustersContainer({ clusterType, projectKey, userPermissions, modifyData, copySnippet });
 };
 
 export default ClustersContainer;
 
 ClustersContainer.propTypes = {
   clusterType: PropTypes.string.isRequired,
+  copySnippet: PropTypes.func,
 };
