+++ code/workspaces/web-app/src/pages/DaskPage.js
@@ -1,16 +1,35 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
+import copy from 'copy-to-clipboard';
 import Page from './Page';
 import ClustersContainer from '../containers/clusters/ClustersContainer';
 import ExternalLink from '../components/common/ExternalLink';
 import { DASK_CLUSTER_TYPE } from '../containers/clusters/clusterTypeName';
+import notify from '../components/common/notify';
 
 const useStyles = makeStyles(theme => ({
   clusterList: {
     marginTop: theme.spacing(5),
   },
 }));
+
+const copySnippet = ({ schedulerAddress }) => {
+  const proxyAddress = schedulerAddress.replace('tcp://', 'proxy/').replace('8786', '8787');
+  const message = `# Paste this into your notebook cell
+# Note that the Dask scheduler can be accessed from the Dask JupyterLab extension with address
+# ${proxyAddress}
+from dask.distributed import Client
+c = Client("${schedulerAddress}")
+c
+`;
+  try {
+    copy(message);
+    notify.success('Clipboard contains snippet for notebook cell');
+  } catch (error) {
+    notify.error('Unable to access clipboard.');
+  }
+};
 
 const DaskPage = () => {
   const classes = useStyles();
@@ -24,7 +43,7 @@ const DaskPage = () => {
         Dask can only be used with Python.
       </Typography>
       <div className={classes.clusterList}>
-        <ClustersContainer clusterType={DASK_CLUSTER_TYPE} />
+        <ClustersContainer clusterType={DASK_CLUSTER_TYPE} copySnippet={copySnippet} />
       </div>
     </Page>
   );
